Add eth_sendTransaction request handling
