+++ Sources/WalletConnect/Utils/Types/EthereumRequest.swift
@@ -1,0 +1,6 @@
+
+import Foundation
+
+public enum EthereumRequest {
+    case ethSendTransaction(EthSendTransaction)
+}
